refactor(core): isolate update supervisor and lifecycle restart helpers
